split header and footer
